+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## [0.2.57](https://github.com/Wildhoney/Chizu/compare/v0.2.56...v0.2.57) (2026-02-11)
+
+### Features
+
+- **hooks:** add context.actions.consume for handler-side broadcast/multicast reads ([3ab2bc5](https://github.com/Wildhoney/Chizu/commit/3ab2bc5475303044fd1da6f47eaa9629d527f3a2))
+
 ## [0.2.56](https://github.com/Wildhoney/Chizu/compare/v0.2.55...v0.2.56) (2026-02-11)
 
 ### Features